Hide replay button if user has no permission
A widget on a dashboard has a play button which directs the
user to the search on which the widget is based on.
If the user had no permission to look on the streams (no read
permission on the stream) he got the message:

'The stream is not available, cannot replay search.'

Some user thought that the button was broken, since they assumed
to trigger a update with this button.

To fix the problem we remove the button for users without
the necessary permissions.

Fixes #4120